gherkin: (C) Manually convert uri:s from char* to wchar_t*.
On Alpine Linux 3.5 running in Docker swprintf does not handle the
format string L"%hs" properly and the written string becomes empty,
therefore swprintf cannot be used to convert the uri:s passed as
argument (as char*) to wide characters (wchar_t*). To work around this
the uri:s are converted manually character by character, taking into
account that the uri is utf8 coded (which only make a difference if the
uri contains non-ascii characters).