datatables: Fix priority of default converters
In some scenarios default converters would be used when a defined
converter was available. This was caused by returning a default
converter as soon as no regular converter was available.

This heuristic works well with two exceptions:
 * When the table contains a single line the default table entry
   transformer is never applicable
 * When the table keys don't imply a table entry the default table
   entry transformer is never applicable

By inlining this lookup and working out the exceptions the
old behaviour is restored.

Fixes: https://github.com/cucumber/cucumber-jvm/issues/1489
Fixes: https://github.com/cucumber/cucumber-jvm/issues/1490