Add stricter gem restrictions for google-protobuf
As the comments elude to; this gem is a key runtime dependency. And on Ruby 2.6 this only works when using the latest version

For Running on JRuby and a couple of other environments, a weaker requirement is required (And not the latest), due to some issues with those environments

Once the team maintaining this has done some additional work, we can probably relax this requirement. But for now this makes cucumber build properly again X-Ruby Versions