cucumber-expressions: Simplify java heuristic
Explaining when a string is interpreted as a cucumber expression should
be rather straight forward. In essence the explanation should be:

    If you surround a string with `^` and `$` it will be a regular
    expression. If not, it is a cucumber expression.

However the heuristic complicates this by checking if a capture group
contains regular expression symbols `[]+.*` with the assumption that
these are not used in normal writing.

This allows us to write little marvels like:

    this looks\( i.e: no regex symbols) like a cukexp
    a heavy storm forecast \(BF {int}+)
    the temperature is (\+){int} degrees celsius

Which look deceptively like cucumber expressions but are in fact regular
expressions. Simplifying the heuristic remove this ambiguity. This does
comes at the cost of being able to recognize obvious regular
expressions.

    this (.+) like a regexp
    this (\d+) like a regexp
    this ([a-z]+) like a regexp

But I think the simplicity is worth it.

Related:
  * https://github.com/cucumber/cucumber-jvm/issues/1581
  * https://github.com/cucumber/cucumber/issues/515