datatable: Reimplement Jackson TypeFactory functionality (#682)
Given an arbitrary type the data table converter creates an instance of
that type from the data in the table. The conversion is handled by
various converters. These converters are registered with the type
convert to. So matching a given target type to the available converters
requires deconstructing the target type to its component parts.

The Jacksons `TypeFactory` provided this functionality for us. However
including Jackson as a (shaded) dependency creates a problem:

 1. To avoid dependency hell Jackson must be shaded in.
 2. Jackson receives frequent security updates. This means DataTable
    needs to updated frequently.

So rather then try to keep up with the frequent releases it would be
better to reimplement part of Jackson functionality.

This implementation provides a simple type system of three types: List,
Map and Other. No other types are needed as our data table
implementation only converts to map, list (of lists) and singletons.