Update Perl modules based on *.feature.token tests
The token tests show that the language token is case-insensitive (tests say
'Language', but the module wanted 'language').

The token tests show that the DocStringSeparator token wants the actual token
'keyword' matched to be included in the token output.

The token tests show that the scenario outline is reported as a keyword match
on the ScenarioLine (whereas it used to be its own token class
ScenarioOutlineLine).