Bind search to URL search param
In the `react` library, the `SearchQuery` can now be optionally bound to
a URL search parameter in the `window.location`. If this option is
enabled, then when the search query context is updated, it causes the
page to reload with an updated search string, rather than updating the
context's state internally.

In `html-formatter`, this new option is used, bound to the 'search'
URL search parameter.

This makes it possible to create hyperlinks to a web page of cucumber
results with a particular search applied. This could be used in DevOps
workflows, for example by attaching a link to an issue ticket that
points to a set of features tagged with the issue number.