gherkin-python: Align the include of default.mk
Align the include of the default.mk in the Makefile of gherkin-python
with tag-expressions-python. Normally in the gherkin implementations
the include of default.mk is done first in the Makefile, but that
assumes that the first make goal in default.mk is the goal that should
be executed by default. Since tag-expressions-python (which created the
default.mk for python) includes the default.mk last, that assumption
does not hold.

Fixes #1481