Convert ruby messages duration.nanos to integer
`duration.nanos` is described in the schema as being an integer. It was already the case with protobuf.

With the new DTOs, `duration.nanos` became a float. When using the json-formatter to convert ndjson to "legacy" json, an error is raised:

```
ERROR: json: cannot unmarshal number 127100.0 into Go struct field Duration.testStepFinished.testStepResult.duration.nanos of type int64
```

Converting nanos to integer should fix this.