[perl] Make sure CHANGELOG.md is included in Perl dists
Since the [Git::GatherDir] plugin for Dist::Zilla *only* gathers
versioned files and the CHANGELOG.md file is *not* versioned, but rather
copied into the project directory upon release, it needs to be explicitly
included in the distribution tar file.

Also, in order to reduce the number of files a developer needs to get
to know, move the ignore patterns from `MANIFEST.SKIP` to `dist.ini`,
eliminating the need to have `MANIFEST.SKIP` at all.