[Perl] Fix out-of-tree tests
With the testdata removed from the directory of each implementation,
the Perl testsuite was adapted to reference outside of the Perl directory
to access a test file. This strategy doesn't work when a dist has been
created and the dist is running its tests outside of the development
tree.

Instead, copy a single simple test file (which needs to be valid, but
doesn't need to be inline with the rest of the testsuite: we're not
using the results, only the success value) into the test tree.